+++ src/main/java/com/shizzy/moneytransfer/service/CardService.java
@@ -9,9 +9,9 @@ import java.util.List;
 public interface CardService {
     ApiResponse<Card> generateCard(GenerateCardRequest cardRequest);
     ApiResponse<List<Card>> findCardByWalletId(Long walletId);
-    ApiResponse<String> lockCard(Integer cardId);
-    ApiResponse<String> unlockCard(Integer cardId);
-    ApiResponse<String> setCardPin(Integer cardId, String pin);
-    boolean checkPin(Integer cardId, String enteredPin);
-    ApiResponse<String> deleteCard(Integer cardId);
+    ApiResponse<String> lockCard(Long cardId);
+    ApiResponse<String> unlockCard(Long cardId);
+    ApiResponse<String> setCardPin(Long cardId, String pin);
+    boolean checkPin(Long cardId, String enteredPin);
+    ApiResponse<String> deleteCard(Long cardId);
 }
